Type System: Add type to dataflow expression
Adds Type fields to all dataflow expression nodes. The type of each
expression node is inferred recursively in lower_expression() in
mir_to_flow.

Dataflow expression evaluation only uses type coercion for LIKE
comparisons and CAST operations. The CAST coercion was slightly changed
to ensure that we are not coercing to Type::Null. This should never
happen as Type::Null is reserved for rare cases where the type for the
value cannot be inferred (e.g., SELECT NULL FROM table).

Dataflow expressions are also used in the view for post read filters.
The type for these filters is given at filter creation. However, this CL
does not yet get the type of the column for some filter expressions, and
instead leaves those as Type::Null. These will be updated in a future
CL, but do not currently pose a problem because we do not enforce type
checking.

Adding type to dataflow expressions also allows us to greatly simplify
the column tracing that occurs to get the type of columns in
controller/schema.rs. The tracing that still happens in schema.rs is
done to resolve the base for columns to build the final column
specification. This can be removed in a future CL as we now store the
source of each column in addition to the type.

There remains a bit of work to do in ensuring the accuracy of the Type
values in each expression node. This CL uses the same logic that was
present in noria to infer type. A follow-up type inference CL is planned
to clean much of this up.

Additionally there is future work to do in building out testing for type
accuracy (compared against mysql and postgres) and type checking.

Change-Id: Icda2d1fb72d8e2e22a79f516a707925b96bb8446
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2205
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>