installer: Support dockerizing with single adapter in standalone mode.
This CL supports an alternate docker installer mode: using ReadySet in
standalone mode. This passes `NORIA_SERVER` to the adapter, creating
the adapter/server as a single container, allowing for reads from the
cache without paying the cost of (de)serialization and a network hop.
The noria-server container is removed, and all modifications on its
compose fields are gated by a flag.

This is currently set via a single constant in the installer script
since this is not something we want users to be able to configure
initially. In the future, this standalone mode will be removed when
we have a complete embedded reader implementation.

Change-Id: I6ab1b91cae8aa16f6534dfa0b48c2039722270a7
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2218
Tested-by: Buildkite CI
Reviewed-by: Peter Farr <peter@readyset.io>