mysql-srv: Send auth switch if client sends no plugin
It seems to be the case that if the mariadb-client (but not the
mysql-client) is run with a database name passed on the command line,
it *doesn't* send the name of its auth plugin as part of its handshake
response packet - in this case, we still need to ask it to send the
correct auth data using the correct plugin. This tweaks the boolean
logic such that that happens correctly, and also commits some debug
logging I added to help diagnose that issue.

Change-Id: I32123a88f9434d7f70f1345056c166e0cdf4548f
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2219
Tested-by: Buildkite CI
Reviewed-by: Justin Miron <justin@readyset.io>