Use vendored OpenSSL for all darwin targets
Previously we would only use vendored OpenSSL for the musl target. We
should do this for all targets we believe our customers may use, to
elminate extra library dependencies for our customers. This commit makes
the change so that we always compile OpenSSL during installer
compilation time for all intended targets to reduce an extra dependency.

Change-Id: Ie66ad365665b900e7ba7e4890a0fea249890bcfc
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2221
Reviewed-by: Griffin Smith <griffin@readyset.io>
Reviewed-by: Justin Miron <justin@readyset.io>
Reviewed-by: Jay Bensal <jay@readyset.io>
Tested-by: Buildkite CI