installer: Add a DeploymentStatus enum
Add a new DeploymentStatus enum, representing the state machine of an
active deployment in the installer. Currently, this just has two
variants, InProgress (the default) and Complete, but later this will be
extended to add a TearingDown status as well. In preparation for adding
tear down support, this also checks if a deployment is already complete
when resuming the installer with an existing deployment, and prints a
message to the user saying so and giving them the URL to connect (this
is factored out from the two other places we were doing this previously)

Refs: ENG-1144
Change-Id: I31bdf7018c73bd2c0ac9a871c0f1cbb8d60fb628
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2225
Tested-by: Buildkite CI
Reviewed-by: Jay Bensal <jay@readyset.io>
Reviewed-by: Peter Farr <peter@readyset.io>