installer: Add entrypoint for teardown
Add a couple of entrypoints for tearing down existing deployments:

1. If the user runs the installer with an existing deployment, give them
   the option to tear that deployment down
2. Add a new "tear-down" (naming tbd) subcommand, which can be specified
   on the command line of the wrapper script invocation like:

       bash -c $(curl -sSL https://launch.readyset.io) tear-down

   This can be optionally passed a deployment name, or it'll prompt for
   the deployment to tear down if none is given

Once we start tearing down a deployment, we update its status to a new
DeploymentStatus::TearingDown status - then if we encounter a deployment
with this status again we continue the tear down process.

The actual tear down process is left unimplemented as of this commit -
both cloudformation and docker-compose just return an error.

Refs: ENG-1144
Change-Id: I144afe7c5bb580929a04cf07fc8ca28f7bf15039
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2235
Tested-by: Buildkite CI
Reviewed-by: Peter Farr <peter@readyset.io>