Add extend_recipe support for serialized SqlQuery
Creates a new extend_recipe_bincode RPC to allow NoriaConnector to send
SqlQuery instead of the &str sql text. RecipeSpecBincode and various
extend_recipe_bincode functions are added to support this.

All extend_recipe calls from NoriaConnector and MigrationHandler now use
the serialized SqlQuery format. The legacy format is maintained for use
by the replicator.

extend_recipe_bincode does not support multiple queries per call and
also does not allow CREATE TABLE statements to create leaves. Support
for both could be added trivially.

This adds a few unfortunate clones in the adapter, which we should
attempt to remove in a future CL.

This change is necessary for overriding the Display impl for
nom_sql::Literal, which the current extend_recipe protocol relies on

Change-Id: I7d74add14c9817d362d487569f159437cc708b08
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2242
Tested-by: Buildkite CI
Reviewed-by: Vlad Krasnov <vlad@readyset.io>