mir: Don't derive hash for column
Having a Hash impl that doesn't match Eq in Rust is bad news - it can
result in all sorts of nasty unexpected behavior when eg storing values
in HashMaps or HashSets, and generally can be quite confusing for
developers. We've *long* had a problem where the Hash and Eq impls for
mir::Column don't match up, because the Eq impls consider "aliases" when
testing if two columns are equal - this hasn't caused us problems yet,
but is a bit of a ticking time bomb, especially if we're about to start
making major changes to the structure of Column.

This commit removes the Hash impl for Column entirely, adjusting all
places where we currently depend on that impl to either use
`nom_sql::Column` in the places where that works, or do slower quadratic
deduplication in the places where we really do want the aliasing
behavior of Column.

I *did* attempt to accomplish all of this by just removing aliases, but
that unfortunately doesn't appear to be practical at the moment

Change-Id: I050ae062c43db090a56cbf0129b5f0f1304f0775
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2250
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>