mir: fully encapsulate Node::columns
In preparation for making `columns` derived on-demand from the inner
node and the node's parent rather than pre-computing and denormalizing
it, refactor MirNode to make the columns *field* private, and only
exposed via a method. This doesn't actually change any behavior (besides
requiring moving a constructor for reuse inside of MirNode) but paves
the way for making this method have a little bit more logic.

Change-Id: Id51da09e7b792fda47b83a23b1a89b70b44c341c
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2252
Reviewed-by: Fran Noriega <fran@readyset.io>
Tested-by: Buildkite CI