sql: Don't take Vec<&Column> anywhere
There are a lot of places in MIR where functions accept `Vec<&Column>`,
only to unconditionally iterate over that vec of references and clone
each - actually basically *everywhere* we accept `Vec<&Column>` we do
this! Constructing that list of references is annoying, and probably
actually slows things down rather than just cloning the vec when we need
to.

Change-Id: I1cfd2a2bceb6cd782fd6879ff391e14f594b1d2f
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2253
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>