Sanitize query logs for potential PII
Removed structs from logs and error messages that might expose PII from
query logs. The removed included queries, expressions, keys, etc.
Created a "display_literals" feature flag to revert to the un-sanitized
behavior.

The following calls to generate errors/logs were sanitized.

- info!()
- warn!()
- error!()
- ReadySetError variants
- internal_err() calls
- internal!()
- invariant!()
- invariant_eq!()
- invariant_ne!()
- all "_span!": info_span!(), debug_span!(), root_span!(),
child_span!(), remote_span!(), quote_spanned!()
- in_scope() for Span
- '#[instrument' calls, including the ReadySet tracing proc macros.
These were not enabled conditionally and should be revisited once we
have a smarter way to sanitize logs

trace!() was not sanitized and exposes significant PII, so we must be
aware of this when setting the log level to trace.

SelectStatements could all be sanitized before being logged, but I chose
not to do this out of concern for performance. Instead, the query is not
printed at all.

We may still have PII exposed through other libraries that I did not
sanitize. The most likely places that this may happen are when we return
errors for mysql_async, tokio_postgres and potentially mysql_srv and
psql_srv. The risk of leaking PII from these libraries is likely very
low, and getting useful errors is probably worth the risk.

Change-Id: Ic4e4c5a2c452a856a4c5ecbc634c4d2e084a0dfe
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2263
Tested-by: Buildkite CI
Reviewed-by: Justin Miron <justin@readyset.io>
Reviewed-by: Griffin Smith <griffin@readyset.io>