mir: Introduce AliasTable node
There are two main reasons we store columns on MIR nodes:

1. To let later nodes know about columns that are produced by those
   nodes' operations, such as the results of project evaluating
   expressions or aggregates' result columns.
2. To note the changed *tables of columns, even those that a node
  doesn't actually change, for the leaf node of a query - so that you
  can eg query from a view or CTE as if its columns have that view's
  name as their table name

This latter case is not going to be possible using our existing set of
MIR nodes once we're deriving columns from the mir node's parent and
node type. To prepare for that, this commit introduces a new AliasTable
MIR node (bikeshed pending) whose only job is to take all of its input
columns and change their table to a given new table, and remove all
column aliases. This node is now used before the leaf of every
query (and as the last node of every view or CTE), *instead* of directly
altering the table on the columns of the last node via the (now deleted)
sanitize_leaf_node function. When converting to dataflow, this node is
skipped entirely, as the only purpose it serves is changing column names
in MIR, and dataflow doesn't have (or, really, doesn't care about)
column names, only indices

Lots of stuff had to be slightly tweaked here to keep reuse sort of
working, but since we're not actually using reuse in production anywhere
at the moment I stopped short of fixing one final reuse-related bug with
a relatively large test (`integration::lobsters`) since that feels like
it'd take too long to fix right now.

Another interesting thing that had to be fixed to get this working is to
make project nodes not return the columns they emit for expressions and
literals in `referenced_columns` (they never really should have been
there in the first place). This was never breaking in the past since
pull_columns just no-ops if it can't find a required column in an
ancestor, but with alias_table we ran into an issue where pull_columns
tried to pull a projected but non-needed `grp` helper for an ungrouped
aggregate *through* an AliasTable node, which obviously broke since the
table name got added.

Change-Id: Id783853323380ac1da271535c523733af9da1f2d
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2259
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>