mir: Derive columns from the node type
Now that we have an AliasTable MIR node, the output columns of a node
can always be derived 100% from the input columns and the fields of the
MirNodeInner denoting the operations that the node performs on its
input. This means that the `columns` field on the Node type is a
denormalization - and one that has caused a lot of issues in the
past (tons of bugs around accidentally getting this denormalized field
mismatched with the output column of the node, plus lots of annoying
manual work to keep the field up-to-date when rewriting MIR graphs like
we do in `decorrelate`). Instead of storing that denormalized field,
this commit makes the `columns()` method just *calculate* the output
columns of the node based on the input columns and the node's Inner
fields.

Some notable changes that were required to get this fully working:

* The Leaf MirNodeInner had to be updated to include the various columns
  referenced in its keys and post-lookup operations in its
  `referred_columns`, so that those columns were properly pulled by the
  `pull_columns` pass - previously those were just put on the columns
  list of the node, even if it didn't project those columns from its
  parent!
* Similarly, to allow those columns to be pulled through the
  `alias_table` node that now exists before every leaf, we had
  to *alias* the table name on those columns, so that they properly
  matched both the source column in the ancestor and the eventual output
  column of the alias_table node itself.
* The decorrelate pass now also needs to do a little bit of expression
  rewriting when pushing filter nodes down below alias_table nodes,
  since otherwise the column table ends up being wrong - to support
  this, this adds a new `referred_columns_mut` method to
  nom_sql::analysis which is basically a complete copy of
  `referred_columns` except that it provides mutable references.

Change-Id: Ia7c178ec78fc339fda3a22093f66c9e5724d094c
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2265
Reviewed-by: Fran Noriega <fran@readyset.io>
Tested-by: Buildkite CI