mir: Print more information in graphviz for Leaf
To aid in debugging issues with column references, print more
information about things like post-lookup operations etc in the graphviz
for the MIR Leaf node, including the index type, the order by if any,
the limit if any, the returned columns if any, and the post-lookup
aggregates if any. The only thing this leaves out is the default row,
primarily because it's likely to be a little bit larger than can
comfortably fit in graphviz, and also because it currently corresponds
1:1 with post-lookup aggregates - if we decide we need it later, eg
because we're doing something more interesting with it, we can add it
pretty easily.

Change-Id: I479d23c032c89c8aad38fdbed47e1a17eae7a45e
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2264
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>