noria-client: Don't log already-handled errors
In mirror_prepare, we explicitly handle errors caused by either
unsupported queries, or views not being found. In explicit migrations
mode, that second one happens *every time* a user prepares a statement
that hasn't been explicitly created via `CREATE CACHE`, and doesn't
actually represent an "error" per se - insofar as it's not an issue with
the system that needs to be fixed.

After spending a couple hours chasing down this error log under the
mistaken assumption that it *was* a real issue that needed to be fixed,
I'm removing the log entirely for errors caused by unsupported, and
converting the log for errors caused by view not found to a warning.

Change-Id: I7e6be6c6f27c2f4be1c9d96555aa4fca54ab865d
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2267
Reviewed-by: Dan Wilbanks <dan@readyset.io>
Tested-by: Buildkite CI