repl: Add new tool for interactively testing db queries
Add a new readyset-repl binary crate that provides a repl which allows
us to run database queries just like the mysql or psql clients, but also
provides the ability to *prepare and execute* queries - this is
something we've previously been unable to do interactively.

Under the hood, this uses the `rustyline` crate to provide an editable,
history-saving input prompt, then uses `database-utils` to connect to
either a mysql or postgresql database to execute queries.

Currently this support is pretty rudimentary - most notably the
parameters passed to a prepared statement are just strings, and we rely
on eg mysql implicitly coercing parameter values- but it's a solid
start!

    Here's an example of what this looks like to use:

    Welcome to the ReadySet REPL!

    Type `help` for help.

    [mysql://root@127.1/test] ❯ prepare select * from t1 where uid = ?;
    Prepared statement id: 0

    [mysql://root@127.1/test] ❯ execute 0 1;

    1 | 4

Also included in this commit is some new stuff in the `debugging` page
of the docs explaining the use of this tool.

Change-Id: Id2816d2eb95504ab9cc918f1613707a1a9ad60ea
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2270
Tested-by: Buildkite CI
Reviewed-by: Justin Miron <justin@readyset.io>