Fix pending state from standalone
We currently have a bug that causes proxied queries to forever stay in a
"pending" state when we deploy using standalone mode (one container for
both adapter and server). We determined that this was the result of
MySQL not coming up fully yet by the time the single server/adapter
binary has time to come up.

We don't actually need the upstream connection at all unless
validate_queries is enabled. This refactors the migration handler to
take an Option for the upstream DB, and to avoid passing an upstream
along at all in the case that we are simply trying to run dry run
migrations in the background to verify which queries we can, or can not
support.

Change-Id: I563f1abc468a5e1bb5be2d08df10283dd30420a6
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2273
Reviewed-by: Justin Miron <justin@readyset.io>
Tested-by: Buildkite CI