Misc Fixes: Outputs Synchronizer and Changelist parsing
The outputs rpc was changed to return 'CreateCacheStatement's instead of
'SelectStatement's. Because of this the outputs synchronizer stopped
updating the adapter's query status cache to allow these queries to be
run against noria without re-issuing a CREATE CACHE statement.

This also fixes a return statement that was erroneously hidden behind
the display_literals flag.

Will add a clustertest for the outputs synchronizer as a follow-up CL
because that will require modifying clustertests to allow for two
adapters.

Change-Id: I2c6b97104c501645fc0d6fbc8de599ae817ebb63
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2278
Tested-by: Buildkite CI
Reviewed-by: Justin Miron <justin@readyset.io>