Implement StreamingIterator for noria-client result
With the previous changes the best way to return results
from a reader map is by cloning the shared Arc pointer
with the result set for each key. It is especially
beneficial to do so when using a local reader, as it
avoids cloning the data entirely.

It is also beneficial for regular readers to avoid performing
the PostLookup step while holding a ReadGuard, and instead
allows us to perform a single PostLookup across multiple keys
after all lookups are done.

Ideally we need an iterator that can hold a set of Arcs, and
iterate over the individual rows of that Arc, however that is
not possible with the regular Iterator trait. To this end
this change implements a StreamingIterator for the results
returned from noria, which can both iterate over Vec<Row> and
over Arc<Vec<Row>> as needed.

This requires changes to the way results are returned from
queries and a bunch of tests, but doesn't not affect any of
the actual logic.

Change-Id: I92f8830a411ad9adddc0a20d697ceb8b2ff9a6fd
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2284
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>