replicators: Lay groundwork for PG DDL replication
Postgres's WAL replication scheme doesn't natively support replicating
DDL - but we can implement DDL replication by having an event trigger
which inserts a description of the DDL change event into a table, then
special-case handling this table in the replicator to convert those rows
into AlterTable replication actions.

This commit lays the groundwork for all of that, adding a step at
replicator startup to create that table and a trigger to just write drop
table and view events for now, and using all of that to handle those
events inside of the postgres connector.

As of this commit the schema for this table is very much subject to
change - most notably, I'm currently writing only an abstract
description of the event to the table and turning it into an actual DDL
statement inside of the replicator, and may decide to have that DDL
statement construction happen in the trigger - most of those decisions
will come once I add support for create and alter table, which are
significantly more complicated due to the need to determine the schema
of the table inside of the trigger.

Refs: ENG-328
Change-Id: I0bc3f69b563aa036fd55a0410f554f6028c1fa06
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2285
Tested-by: Buildkite CI
Reviewed-by: Vlad Krasnov <vlad@readyset.io>