Create Sensitive and SensitiveDebug to override display
Creates a wrapper struct that overrides the underlying display according
to whether the "display_literals" flag is set. This cleans up the
previous implementation for redacting sensitive information as the
display_literals flag now only appears in launchpad.

The wrapper should be used in any errors or logs that might expose PII.

This CL also permits logging of SET statements, which may leak PII but
are very unlikely to. The benefit of logging SET statements is likely
worth the risk.

For convenience, the display_literals feature is added to noria-mysql,
noria-psql, and noria/server which each pass the feature to launchpad.
This is done to allow conditional compilation for testing by adding a
flag.

Change-Id: I86a9a818d6d6c9afaf60987abef3dc3b76667556
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2288
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>