sql: Add AliasTable after reuse, rather than before leaf
Queries can internally have multiple columns with the same name, for
example a join that projects an "id" column from one side and keys on an
"id" column from the other side of the join. The AliasTable node that we
were adding right before the leaf of the query, however, discarded all
information about which of those duplicate columns came from which
table, which could lead to us incorrectly using the wrong column as eg
the key column of a leaf in that case.

The only reason we have to add an AliasTable node in the first place,
however, is so that the columns returned by a view have the name of the
view as their table, allowing subsequent queries to query from that view
as if it were just a regular base table. This means that we can avoid
the information-loss issue with the table name of the column by adding
the AliasTable node *after* the reuse node we create at the base of the
new query (including subqueries, since those also have the table name
aliased), rather than before the leaf node. Doing this also allows us to
avoid calling `aliased_as_table` on all the columns referenced by a leaf
node, as they no longer need to be able to be projected *through*
AliasTable nodes.

This mostly works fine right out of the box, except in the case of query
reuse, where something strange I don't understand is causing an
assertion to fail that previously wasn't - just removing this assertion
made everything work, however, so while I'm deeply suspicious, I'm just
leaving that as a debug log for now as reuse is always disabled in
production.

This commit also includes the logictest that originally reproduced this
issue, since otherwise it's only found by a vertical test (yikes!).

Change-Id: Ia8c7ad9bc543fe14d3bf03b80d365f7ed52fd049
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2289
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>