nom-sql: Expose better wrappers for specific parsers
Rather than exposing internal parser combinators for parsing specific
kinds of statements like SelectStatement and AlterTableStatement, it's
more convenient and more correct (due to not having to remember to check
for trailing input after a successful parse, and not having to convert
the input to bytes) to expose specific parser wrappers in the style of
`parse_query` etc. This does that for the specific things we had been
exposing directly (selection, creation, and alter_table_statement) and
updates the call sites to use those instead.

Change-Id: I65f066403876ce900378c1b958cf05fb53635fce
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2295
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>