replicators: replicate CREATE TABLE for PG
Add a new event trigger definition to the postgresql ddl replication
code which, after a CREATE TABLE event, uses a couple of queries into
pg_catalog tables (mostly equivalent to what we already do for the
initial snapshot) to reconstruct a semblance of the original CREATE
TABLE statement - then, handle that specifically in the ddl replication
event handler, by doing the same "parse as postgresql, then stringify as
mysql" juggle that we do in snapshotting.

Since this is all pretty complex, this comes with a couple of unit tests
for the actual statement construction logic in `ddl_replication.rs`, in
addition to an integration test in the usual place.

Refs: ENG-328
Change-Id: I743eda2d1bf59f1fa3e65b10db5251a1777349f2
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2296
Tested-by: Buildkite CI
Reviewed-by: Vlad Krasnov <vlad@readyset.io>