replicators: Replicate CREATE VIEW for PG
Fully replicate the CREATE VIEW statement for postgres using the
event-trigger based replication scheme we're using for other DDL.

This is implemented by storing the full CREATE VIEW statement in a
column in the ddl_replication_log, which has been changed from
`create_table_ddl` to just `statement` since it's now becoming more
multi-purpose. I've opted to *not* do the same thing for drop table and
drop view, primarily because it allows us to avoid the
parse-then-re-stringify ceremony that we have to do when the native
format of the replication log has the wrong kind of SQL syntax - this is
a little bit of inconsistency, though, so we may want to change our
minds about that at some point.

Refs: ENG-328
Change-Id: I8d50ae994f368714c974db2608717731b169a96a
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2300
Tested-by: Buildkite CI
Reviewed-by: Vlad Krasnov <vlad@readyset.io>