replicators: Replicate ALTER TABLE as CREATE TABLE for PG
Using a postgresql event trigger to reconstruct an actual ALTER TABLE
statement doesn't seem right now like it'd be possible - the trigger
only runs after the table has been altered, so we can't do the thing we
do for create table statements and query the catalog for the structure
of the table. Fortunately, the way we're planning on implementing ALTER
TABLE generally doesn't actually *care* about the actual contents of the
ALTER TABLE statement, only that we receive a notification that the
table *has been* altered. The plan here, then, is to just send a CREATE
TABLE statement to extend_recipe when we get an alter table event in
postgres, then teach the recipe that if it gets a create table statement
for a table it already knows about, it should treat that as an alter
table statement (morally, it's "upsert"-ish).

This commit implements the replicator half of that, including a unit
test making sure that the correct stuff gets written to the
`ddl_replication_log` table, and adds a currently-ignored integration
test to replicators while full fledged alter table support is getting
added on the noria side of things.

Refs: ENG-328
Change-Id: Ia768252437ed37ed2eeff6a90a72cfaa62073822
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2302
Tested-by: Buildkite CI
Reviewed-by: Fran Noriega <fran@readyset.io>