installer: Enable server for pg's adapter, fix noisy healthcheck
This enables the internal rs server in the adapter for the installer's
postgres standalone profile. Without this, things like create table and
probably most other things will just hang the client's connection.

Additionally, for the Postgres healthcheck, instructed it to use the
postgres user (and dropped the localhost argument). This doesn't change
the behavior of the healthcheck, it just prevents Postgres's log from
being spammed with messages about the root user not existing on each
healthcheck.

Change-Id: I93244af19f9c46fe4bfb6dc37766f2538087db43
Refs: ENG-1224
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2325
Tested-by: Buildkite CI
Reviewed-by: Justin Miron <justin@readyset.io>