readyset/helm: introduce components for v0.0.1 of ReadySet Helm Chart
The purpose of this CL is to add a Helm chart for deploying:

- Prometheus
- Grafana
- Consul Server
- Vector (agent, aggregator, etc.)
- ReadySet Adaptor
- ReadySet Server

Currently, this chart should only be deployed with a single server node,
given that inter-domain communication requires a very broad range of TCP
ports which wouldn't be feasible to implement in k8s (for now). We will
look into limiting the scope of those ports in the future, but for now,
the readyset-server StatefulSet should be set to a single replica.

It's worth noting that this initial implementation is to get us up and
running with something, and we can iterate as we go forward.

It's worth calling out that the Vector agent is being run as a sidecar.
Logs are consumed directly from disk and shipped accordingly.

You might notice that a handful of our Grafana dashboards are being
stored in values.yaml. This works for now, but as we move towards a
proper Helm chart release pipeline or process, we may want to
dynamically build ConfigMaps key/values or values.yaml entries based on
a whitelist of dashboards. We can chat more about this later, but wanted
to briefly describe some of the temporary decisions made here.

Mantras/ideas influencing the development of this chart:

 "Make the helm chart as plug-and-play as possible."
 "Minimal values configuration should be needed to deploy the chart."

Change-Id: I8042bbf6b59b0a28866757dc6ba8029db498e0ed
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2327
Tested-by: Buildkite CI
Reviewed-by: Harley Klein <harley@readyset.io>
Reviewed-by: Sasha Fox <sasha@readyset.io>