Account for compose installed with docker
Previously we would only try to find/use a standalone docker-compose. As
of today docker compose is a subcommand of docker, and gets installed
with docker. We shouldn't error out if someone doesn't have
docker-compose, but does have docker compose accessible via the compose
docker subcommand.

This commit prefers docker-compose, but if not found will use `docker
compose` instead.

Change-Id: I74d915f623f818407218a9c507587244d22125db
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2330
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>