graphviz: Render domains as clustered subgraphs
Use graphviz's "cluster subgraph" feature to render domains as labeled
gray boxes containing nodes within that domain. This can make the layout
a bit worse in certain cases, but the benefit of having domain index
labeled on the graph is significant.

To make this work, I had to add an extra argument to `graphviz`
specifying the mapping from domains to nodes that's stored in the
controller - since graphviz is currently called from a few places where
we don't have access to that mapping, I've made this an Option, and if
it's not specified the previous layout will be used.

This commit also includes updates to the docs for graphviz to make the
example use this new layout, and add a bullet point to the "interpreting
graphviz" section explaining what's going on here.

Change-Id: I90ea664535e5d13175fa3c5bacc69781af46993a
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2333
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>