Ensure we wait for mysql/postgres healthcheck
Previously with the non-standalone version of deployment,
readyset-adapter depended on readyset-server being healthy, which in
turn depended on mysql/postgres being healthy. When the server got
removed in favor of a standalone versino of the adapter/server, the
depends on relationship on the server got removed. This means that we
have nothing which waits to see if mysql is first healthy before
proceeding. This can easily lead to a situation where we tell a user
their readyset cluster is (locally) deployed, but it's not actually
available yet.

By adding the depends on relationship for mysql and postgres
respectively, a user should have a better experience, by being able to
successfully connect after we provide a connection string.

Change-Id: I9069f0800e7b57ea627f29605bd201fa9ceb6a40
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2335
Reviewed-by: Corbin Hall <sif@readyset.io>
Tested-by: Buildkite CI