Add API key protections
We desire for a user to be required to input a provided key, as a small
level of protection. The security requirements for this feature are
extremely low, so baking the API key into the binary is considered
acceptable.

If better security is desired in the future, the solution should be
changed.

Change-Id: I6d55b6adc4e1f36830b0a8b1851de5e229257392
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2337
Tested-by: Buildkite CI
Reviewed-by: Jay Bensal <jay@readyset.io>
Reviewed-by: Dan Wilbanks <dan@readyset.io>