mysql-srv: send auth switch request if password is empty
It looks like some versions of the php pdo mysql connector send an empty
password response in their initial handshake, even if the auth plugin is
present and correct. This *seems* to me to be incorrect behavior per the
mysql binary protocol spec, but just sending an auth switch request if
it happens causes the issue to go away.

Refs: ENG-1237
Change-Id: I2b5d2741aa4479a51a24a47e5c8e8887dbefcb70
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2338
Reviewed-by: Michael Cronce <mc@readyset.io>
Tested-by: Buildkite CI