installer: Fix checking for docker+compose installation
If the program for a Command does not exist, actually spawning that
command fails with an io::ErrorKind::NotFound error - but we were only
checking the *status* of the command, which can only be set if the
command was actually run in the first place. This meant that running the
installer on a machine without docker-compose installed would just fail
with an "Error: No such file or directory (os error 2)" instead of
giving an actually helpful message indicating that docker or
docker-compose wasn't installed.

This commit fixes that by actually checking for the NotFound error and
giving a more useful error message, doing so in a helper function since
we need to do it three times (once for docker, and once for each of the
different possible ways we can run docker-compose). This also needs to
happen for when we actually *run* docker-compose, since we first try
running `docker-compose` and then try running `docker compose`, and the
same issue applies there.

Fixes: ENG-1230
Change-Id: I796f0108b57b6125e8609ebf32179830f14d5e5f
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2344
Tested-by: Buildkite CI
Reviewed-by: Peter Farr <peter@readyset.io>