buildkite: Add installer deploy step
Add a step to the buildkite pipeline to deploy a new release of the
installer to the S3 bucket backing https://launch.readyset.io. This
implements the last of the manual steps that're currently documented in
http://docs/release-process.html - notably the uploading of the binaries
to S3, then the creation and waiting for completion of the cloudfront
invalidation.

This step is still just gated behind the "build release :docker:
containers" block step, since right now we're only really working with
the docker-compose version of the installer, and the cloudformation
version is more on the backburner - once we bring back the
cloudformation version, we likely want to combine the docker and cfn
release steps into one and have this depend on both, since *technically*
the installer is coupled to a specific version of the cloudformation
templates.

Regardless, as of this commit it's now possible to press a single button
to deploy the installer! Yay!

An example of this working in isolation is at
https://buildkite.com/readyset/readyset/builds/16900

Fixes: ENG-1101
Change-Id: I883856d44fe2e61771fa758295c796dd5a8198ab
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2345
Tested-by: Buildkite CI
Reviewed-by: Ryan Cerniglia <ryan@readyset.io>