Orchestrator: (Compose) Don't reuse anon volumes, remove orphans
This changes the `docker compose up` to, additionally, tell it to not
reuse anon volumes, which resolves the issue of creating a new
deployment with the same DB type, but different password, resulting in
auth failures issue. On top of that, it also instructs it to remove
oprhaned containers, which will resolve the issue where switching
between mysql<->postgres results in the previous database sticking
around.

Refs: ENG-1259
Change-Id: I0846d5d084cb817cd87ccbefe7b36423c77b04f5
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2348
Tested-by: Buildkite CI
Reviewed-by: Griffin Smith <griffin@readyset.io>