Grant installer role perms to create cloudfront invalidations
One of the steps we need to do as part of the automated installer
release process is to create an invalidation for the cloudfront
distribution serving launch.readyset.io. This grants permission to do
that to the InstallerS3 role that already exists with permissions to
upload to the installer bucket itself - unfortunately, we have to
hardcode the ARN of the distribution itself here as it's created in
another root module and seemingly there's no way to pass resources like
this between root modules in substrate's module.
72cec33fe (Modifications to Installer Bucket, 2022-04-13) had to do the
same thing in the opposite direction, so this is at least somewhat
precedented, if annoying.

Change-Id: I155b29ac76a01286e6dea256cf1bae96d36e4f1f
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2353
Tested-by: Buildkite CI
Reviewed-by: Harley Klein <harley@readyset.io>
Terraform-Plan-Review: Harley Klein <harley@readyset.io>