adapter: Log warning on startup for allow-unsupported-set
Blindly allowing unsupported set statements in the worst case has the
potential to cause queries to return *drastically* incorrect results -
for example if the ANSI_QUOTES sql mode is set, string literals could be
incorrectly interpreted as references to columns, which is pretty
severe. Even though this flag is hidden, I'm electing to err on the side
of extreme caution here and *also* log a warning at startup if the flag
is enabled.

Change-Id: I96f818b597e8406ac6ba81a58395cd6e2fa56ddc
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/2355
Tested-by: Buildkite CI
Reviewed-by: Dan Wilbanks <dan@readyset.io>